Several things going on in this submission. First, there's been a long time lag from the last submission and some files were left with uncommitted changes. Also, since the last commit it was determined over coffee that we needed function calls and parameters. Those parameters had to allow pass-by-value and pass-by-reference. This is an incomplete attempt to capture those requirements.
Tried to capture the differences between function arguments in the definition and at the call stage. Calling a function can have another function call as a parameter, but the function declaration can only have a variable name with an optional pass-by-reference symbol. In this case we've chosen to use & like in C++. The pointer * is not used in this language.

Because of the changes here, the interpreter will also require substantial upgrading.

Also, bitwise AND and OR were added to the binary operator table. Don't know how those got left out but adding pass-by-reference, which uses the & symbol, made it apparent that we needed this symbol for bitwise AND too.

Have tried to be more strict in the use of constants in function declarations. However, since this is not C++, it's value is limited.

Also tried to encapsulate the use of file pointers to where they are used. Previously they were global.