Adding code to allow the embedded motor controller to accept statement and symbol entries over a serial line and add them to the interpreter's internal state. The way this works is that after each character is received, the interpreter checks to see if it contains a valid line, if it does then it stores it in its internal table; if not then it just waits for the next character.
Note that tinyQueue.h was modified so that the constructor can specify where the first and last pointers are in the buffer. This was done so the incoming string of bytes is stored going left to right. The default implementation would have done this backwards. Normally you wouldn't have to worry about the internal workings of this, but you can now request the buffer which is then passed to the parser. Also, this is implemented as a circular queue but we never wrap the indices. In other words, if we can't find a valid line to parse by the time the buffer fills up then something is wrong. The only index in the buffer that we change if _front.