This message duplicated from motor repository.
Saving and restoring program to flash memory tested for first 3 entries in the interpret::_program[] table. See matching submission in parser repository.
Current thinking is to add a button to the board that would transfer the program from flash to sRAM and begin executing.
Another thought is to keep the program table, but not the symbol table (since it changes values), in flash all the time. That is a future optimization.
No testing done on symbol table, but that should be straightforard.
Note that one entry in _program[] takes up 12 bytes. In _symbolTable[] it's 16 bytes.